requirements.txt
anyio==3.5.0
asgiref
certifi==2021.10.8
click==8.0.4
dlib==19.23.0
docopt==0.6.2
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.74.1
h11
idna==3.3
joblib==1.1.0
numpy==1.22.2
opencv-python==4.5.5.62
Pillow==9.0.1
pydantic==1.9.0
scikit-learn==1.0.2
scipy==1.8.0
sniffio==1.2.0
starlette==0.17.1
threadpoolctl==3.1.0
typing_extensions==4.1.1
uvicorn
